// src/screens/Home/enums.ts
/* eslint-disable no-unused-vars */
enum LanguagesEnum {
  ES = 'ES',
  EN = 'EN',
  PR = 'PR',
}

enum ColorsEnum {
  RED = 'RED',
  BLUE = 'BLUE',
  GREEN = 'GREEN',
  YELLOW = 'YELLOW',
}

enum ThemeEnum {
  COLORS = 'COLORS',
  NUMBERS = 'NUMBERS',
  ANIMALS = 'ANIMALS',
}

export { LanguagesEnum, ColorsEnum, ThemeEnum };